Fix inner circle's border width

+++ app/src/main/java/com/antonpopoff/standstatsgenerator/views/ViewExtensions.kt
@@ -2,4 +2,4 @@ package com.antonpopoff.standstatsgenerator.views
 
 import android.view.View
 
-fun View.pxToDp(px: Int) = this.resources.displayMetrics.density * px
+fun View.pxToDp(px: Float) = this.resources.displayMetrics.density * px
